requirements.txt
setuptools 
wheel 
torch 
torchvision 
torchaudio
pyaudio 
Cython
soundfile 
librosa
pyttsx3
openai-whisper
requests
gtts
TTS

ddgs
aiohttp
trafilatura
streamlit

langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-chroma
langchain-huggingface
langchain-ollama
langgraph
langchain-openai

beautifulsoup4

playwright
ollama
google-genai

pypdf
docling
onnxruntime
fastapi
uvicorn

pypdf
chromadb
sentence-transformers


pytesseract
pillow
tesserocr

numpy>=1.24.4,<2.0.0
opencv-python-headless<4.11
opencv-python<4.11